+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.1] - 2025-08-05
+### Added
+- Save database errors to *.err file.
+
+### Fixed
+- Don't break on append error.
+- Don't generate the *.err file if there's no errors.
+
 ## [0.1.0] - 2025-08-05
 First version
 
+[0.1.1]: https://github.com/oscarotero/log2duck/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/oscarotero/log2duck/releases/tag/v0.1.0